+++ mlrun/model_monitoring/model_monitoring_batch.py
@@ -28,6 +28,7 @@ import v3io.dataplane
 import v3io_frames
 
 import mlrun
+import mlrun.common.model_monitoring
 import mlrun.common.schemas
 import mlrun.data_types.infer
 import mlrun.feature_store as fstore
@@ -538,7 +539,7 @@ class BatchProcessor:
 
         # Get the batch interval range
         self.batch_dict = context.parameters[
-            mlrun.model_monitoring.EventFieldType.BATCH_INTERVALS_DICT
+            mlrun.common.model_monitoring.EventFieldType.BATCH_INTERVALS_DICT
         ]
 
         # TODO: This will be removed in 1.5.0 once the job params can be parsed with different types
@@ -554,7 +555,8 @@ class BatchProcessor:
 
         # Define the required paths for the project objects
         tsdb_path = mlrun.mlconf.get_model_monitoring_file_target_path(
-            project=self.project, kind=mlrun.model_monitoring.FileTargetKind.EVENTS
+            project=self.project,
+            kind=mlrun.common.model_monitoring.FileTargetKind.EVENTS,
         )
         (
             _,
@@ -564,7 +566,7 @@ class BatchProcessor:
         # stream_path = template.format(project=self.project, kind="log_stream")
         stream_path = mlrun.mlconf.get_model_monitoring_file_target_path(
             project=self.project,
-            kind=mlrun.model_monitoring.FileTargetKind.LOG_STREAM,
+            kind=mlrun.common.model_monitoring.FileTargetKind.LOG_STREAM,
         )
         (
             _,
@@ -617,18 +619,24 @@ class BatchProcessor:
 
         for endpoint in endpoints:
             if (
-                endpoint[mlrun.model_monitoring.EventFieldType.ACTIVE]
-                and endpoint[mlrun.model_monitoring.EventFieldType.MONITORING_MODE]
-                == mlrun.model_monitoring.ModelMonitoringMode.enabled.value
+                endpoint[mlrun.common.model_monitoring.EventFieldType.ACTIVE]
+                and endpoint[
+                    mlrun.common.model_monitoring.EventFieldType.MONITORING_MODE
+                ]
+                == mlrun.common.model_monitoring.ModelMonitoringMode.enabled.value
             ):
                 # Skip router endpoint:
                 if (
-                    int(endpoint[mlrun.model_monitoring.EventFieldType.ENDPOINT_TYPE])
-                    == mlrun.model_monitoring.EndpointType.ROUTER
+                    int(
+                        endpoint[
+                            mlrun.common.model_monitoring.EventFieldType.ENDPOINT_TYPE
+                        ]
+                    )
+                    == mlrun.common.model_monitoring.EndpointType.ROUTER
                 ):
                     # Router endpoint has no feature stats
                     logger.info(
-                        f"{endpoint[mlrun.model_monitoring.EventFieldType.UID]} is router skipping"
+                        f"{endpoint[mlrun.common.model_monitoring.EventFieldType.UID]} is router skipping"
                     )
                     continue
                 self.update_drift_metrics(endpoint=endpoint)
@@ -642,12 +650,12 @@ class BatchProcessor:
                 _,
                 _,
             ) = mlrun.utils.helpers.parse_versioned_object_uri(
-                endpoint[mlrun.model_monitoring.EventFieldType.FUNCTION_URI]
+                endpoint[mlrun.common.model_monitoring.EventFieldType.FUNCTION_URI]
             )
 
-            model_name = endpoint[mlrun.model_monitoring.EventFieldType.MODEL].replace(
-                ":", "-"
-            )
+            model_name = endpoint[
+                mlrun.common.model_monitoring.EventFieldType.MODEL
+            ].replace(":", "-")
 
             m_fs = fstore.get_feature_set(
                 f"store://feature-sets/{self.project}/monitoring-{serving_function_name}-{model_name}"
@@ -660,14 +668,16 @@ class BatchProcessor:
                 df = m_fs.to_dataframe(
                     start_time=start_time,
                     end_time=end_time,
-                    time_column=mlrun.model_monitoring.EventFieldType.TIMESTAMP,
+                    time_column=mlrun.common.model_monitoring.EventFieldType.TIMESTAMP,
                 )
 
                 if len(df) == 0:
                     logger.warn(
                         "Not enough model events since the beginning of the batch interval",
                         parquet_target=m_fs.status.targets[0].path,
-                        endpoint=endpoint[mlrun.model_monitoring.EventFieldType.UID],
+                        endpoint=endpoint[
+                            mlrun.common.model_monitoring.EventFieldType.UID
+                        ],
                         min_rqeuired_events=mlrun.mlconf.model_endpoint_monitoring.parquet_batching_max_events,
                         start_time=str(
                             datetime.datetime.now() - datetime.timedelta(hours=1)
@@ -684,7 +694,7 @@ class BatchProcessor:
                 logger.warn(
                     "Parquet not found, probably due to not enough model events",
                     parquet_target=m_fs.status.targets[0].path,
-                    endpoint=endpoint[mlrun.model_monitoring.EventFieldType.UID],
+                    endpoint=endpoint[mlrun.common.model_monitoring.EventFieldType.UID],
                     min_rqeuired_events=mlrun.mlconf.model_endpoint_monitoring.parquet_batching_max_events,
                 )
                 return
@@ -696,13 +706,15 @@ class BatchProcessor:
 
             # Create DataFrame based on the input features
             stats_columns = [
-                mlrun.model_monitoring.EventFieldType.TIMESTAMP,
+                mlrun.common.model_monitoring.EventFieldType.TIMESTAMP,
                 *feature_names,
             ]
 
             # Add label names if provided
-            if endpoint[mlrun.model_monitoring.EventFieldType.LABEL_NAMES]:
-                labels = endpoint[mlrun.model_monitoring.EventFieldType.LABEL_NAMES]
+            if endpoint[mlrun.common.model_monitoring.EventFieldType.LABEL_NAMES]:
+                labels = endpoint[
+                    mlrun.common.model_monitoring.EventFieldType.LABEL_NAMES
+                ]
                 if isinstance(labels, str):
                     labels = json.loads(labels)
                 stats_columns.extend(labels)
@@ -719,11 +731,13 @@ class BatchProcessor:
             m_fs.save()
 
             # Get the timestamp of the latest request:
-            timestamp = df[mlrun.model_monitoring.EventFieldType.TIMESTAMP].iloc[-1]
+            timestamp = df[mlrun.common.model_monitoring.EventFieldType.TIMESTAMP].iloc[
+                -1
+            ]
 
             # Get the feature stats from the model endpoint for reference data
             feature_stats = json.loads(
-                endpoint[mlrun.model_monitoring.EventFieldType.FEATURE_STATS]
+                endpoint[mlrun.common.model_monitoring.EventFieldType.FEATURE_STATS]
             )
 
             # Get the current stats:
@@ -744,7 +758,7 @@ class BatchProcessor:
             monitor_configuration = (
                 json.loads(
                     endpoint[
-                        mlrun.model_monitoring.EventFieldType.MONITOR_CONFIGURATION
+                        mlrun.common.model_monitoring.EventFieldType.MONITOR_CONFIGURATION
                     ]
                 )
                 or {}
@@ -764,7 +778,7 @@ class BatchProcessor:
             )
             logger.info(
                 "Drift status",
-                endpoint_id=endpoint[mlrun.model_monitoring.EventFieldType.UID],
+                endpoint_id=endpoint[mlrun.common.model_monitoring.EventFieldType.UID],
                 drift_status=drift_status.value,
                 drift_measure=drift_measure,
             )
@@ -776,14 +790,16 @@ class BatchProcessor:
             }
 
             self.db.update_model_endpoint(
-                endpoint_id=endpoint[mlrun.model_monitoring.EventFieldType.UID],
+                endpoint_id=endpoint[mlrun.common.model_monitoring.EventFieldType.UID],
                 attributes=attributes,
             )
 
             if not mlrun.mlconf.is_ce_mode():
                 # Update drift results in TSDB
                 self._update_drift_in_input_stream(
-                    endpoint_id=endpoint[mlrun.model_monitoring.EventFieldType.UID],
+                    endpoint_id=endpoint[
+                        mlrun.common.model_monitoring.EventFieldType.UID
+                    ],
                     drift_status=drift_status,
                     drift_measure=drift_measure,
                     drift_result=drift_result,
@@ -791,21 +807,23 @@ class BatchProcessor:
                 )
                 logger.info(
                     "Done updating drift measures",
-                    endpoint_id=endpoint[mlrun.model_monitoring.EventFieldType.UID],
+                    endpoint_id=endpoint[
+                        mlrun.common.model_monitoring.EventFieldType.UID
+                    ],
                 )
 
         except Exception as e:
             logger.error(
-                f"Exception for endpoint {endpoint[mlrun.model_monitoring.EventFieldType.UID]}"
+                f"Exception for endpoint {endpoint[mlrun.common.model_monitoring.EventFieldType.UID]}"
             )
             self.exception = e
 
     def _get_interval_range(self) -> Tuple[datetime.datetime, datetime.datetime]:
         """Getting batch interval time range"""
         minutes, hours, days = (
-            self.batch_dict[mlrun.model_monitoring.EventFieldType.MINUTES],
-            self.batch_dict[mlrun.model_monitoring.EventFieldType.HOURS],
-            self.batch_dict[mlrun.model_monitoring.EventFieldType.DAYS],
+            self.batch_dict[mlrun.common.model_monitoring.EventFieldType.MINUTES],
+            self.batch_dict[mlrun.common.model_monitoring.EventFieldType.HOURS],
+            self.batch_dict[mlrun.common.model_monitoring.EventFieldType.DAYS],
         )
         start_time = datetime.datetime.now() - datetime.timedelta(
             minutes=minutes, hours=hours, days=days
@@ -870,7 +888,7 @@ class BatchProcessor:
             "endpoint_id": endpoint_id,
             "timestamp": pd.to_datetime(
                 timestamp,
-                format=mlrun.model_monitoring.EventFieldType.TIME_FORMAT,
+                format=mlrun.common.model_monitoring.EventFieldType.TIME_FORMAT,
             ),
             "record_type": "drift_measures",
             "tvd_mean": drift_result["tvd_mean"],
