+++ mlrun/model_monitoring/stores/models/mysql.py
@@ -17,7 +17,7 @@
 import sqlalchemy.dialects
 from sqlalchemy import Boolean, Column, Integer, String, Text
 
-import mlrun.model_monitoring.constants as model_monitoring_constants
+import mlrun.common.model_monitoring as model_monitoring_constants
 from mlrun.utils.db import BaseModel
 
 from .base import Base
