+++ mlrun/runtimes/package/context_handler.py
@@ -30,7 +30,7 @@ from mlrun.execution import MLClientCtx
 from mlrun.utils import logger
 
 
-# TODO: Move the `ArtifactType` to constants.py
+# TODO: Move the `ArtifactType` to common/constants/model_monitoring.py
 class ArtifactType(Enum):
     """
     Possible artifact types to log using the MLRun `context` decorator.
