+++ mlrun/serving/server.py
@@ -24,9 +24,9 @@ from typing import Optional, Union
 
 import mlrun
 import mlrun.utils.model_monitoring
+from mlrun.common.model_monitoring import FileTargetKind
 from mlrun.config import config
 from mlrun.errors import err_to_str
-from mlrun.model_monitoring import FileTargetKind
 from mlrun.secrets import SecretsStore
 
 from ..datastore import get_stream_pusher
